requirements.txt
isodate
formant==1.61.*
requests
asyncio
websockets
sav.channels
python-socketio
aiohttp